Add php-fpm docker image
We'll be building a symfony app which will run via nginx and
php-fpm. This starts us off by building a php-fpm image that is
designed to run symfony. It includes any relevant extensions
required, as well as recommendations such as opcache.

A cron variant is also included to we can easily enabled scheduled
jobs if we need.

Also a dev variant is included for the main server and the cron
variant. This is pretty much identical, except it has a slightly
different entrypoint which ensures file permissions (although this
is more of an issue kubernetes than docker). It also slightly
changes some configuration in the php.ini.