Initialise the symfony app
Initialised a base symfony app in the api directory. We'll be using
a multi-tiered architecture (standalone frontend app communicating
with backend api), and we'll be using a mono-repo structure for
keeping everything in one place.

It is a base symfony installation with a new controller in the
./src/Connections/Controller/... directory which is a simple
hello-world endpoint, which we can use to ensure our docker
environment is working.