Setup unit tests
Setup the api to have unit tests available. Basically installed
phpunit and configured two suites, all and unit which run either
all tests or only unit tests. Currently we just have the one unit
test, but this will be useful later.

Added a make task to run the unit tests.

Updated the php-fpm docker image with the zip extension as we
were getting warnings from composer as it was missing. It also got
a hell of a lot quicker when i added it.

I added 1 unit test for the Default controller for the connections
API...just to check it's all working properly.