Add phpcs fixer
Php cs fixer does not currently support symfony 6, so I found a
workaround on a corresponding issue in github. This is documented
in the readme.

Php cs fixer has been installed in tools/php-cs-fixer subdirectory
of the api. This is reccommended way to install according to the
github readme.

Also ran cs fixer on the existing code.

Added make tasks for linting and fixing of coding standards for
the api. Also added one to configure the git hooks in this repo
to run.

Added a pre-commit hook that runs the linter.