Setup DTO resolution
Configured the app so that controllers can typehint a DTO which
will be resolved from the request. This includes adding a custom
resolver which will check any for a specific interface, and if it
matches will construct a dto by passing the current request to the
dto::fromRequest method.

This should make it easy to keep controller methods short,
essentially they will just pass the dto to a command, which will
in turn be handled by it's handler (as per command pattern).

This keeps all of the complexity outside of controllers and in the
domain of the app, rather than in the http layer (controllers).