Add command bus
This is prototypical and i'll be experimenting with a couple of ways.
This seems to work, but requires a fair amount of manual configuration
whereas if we switch to the symfony messenger component we can
use middleware easily which is something we'll want.

This currently sets up a command bus that has a service locator that
will determine how to handle each command it is passed.

Along with this I installed the fos rest bundle and jms serializer
to make it easier to convert the request objects to a dto.