Setup command bus properly
Okay, so bit of a rager going on here...

I switched over to the symfony messenger component instead of the
custom built command bus. This seemed alot simpler to get going and
required alot less bootstrapping type code in services.yaml. Plus
it has the option of middleware, which we'll use to handle
database transactions later.

Added alot of stuff around the conversion of a command and its
handler to something that can be returned from an API. Basically
it was just trying to find the right separation so that the command
and handler are not tied to http, so they can be used via other
entrypoints (i.e. bulk insert, config file etc).

I also created the base model for connections which later be
coverted to an actual doctrine entity when the db layer is added.
A connections list was added to make working with multiple
connections easier, basically its a workaround as you can't
typehint []Connection or something similar; having an object
encapsulate this should help keep the code robust.

I removed the DTO's as the commands were basically just passing
method calls on the DTO's so they seemed unnecessary. All the DTO's
became commands instead.

Alot of tests are broken, and even more haven't been added, but
it's sunday night so...