Help IDE figure out stamp
IDE was kicking off because it could find the getResult method from
the dispatch of a message on the bus. Normally i wouldn't code to
satisfy the IDE, but in this case there is little reason not to,
a variable declaration maks it easier to read as well as for the
IDE to figure out.