Add validation for commands
Started off by including the validator component for symfony.

Updated field for the list connections command to more accurately
reflect what will be expected. Added validation attributes to all
the fields.

Added an Error namespace within Common/API to encapsulate some
classes that help with creating a consistent representation for
errors across the API.

Included errors within the json response builder.

Moved the RequiredMessageBus trait to a DI namespace within Common
as we now have a couple of these types of traits with the addition
of RequiresValidator. These will be stored in the DI namespace.

Added a new trait to help convert symfony validation results into
the more domain specific version of validation errors which is
easier to convert to a predictable response in the API responses.

Added a create connections response and response builder so that
it actually works properly now, rather than just returning a string.

Updated command handlers to actually validate the command. This
mostly works correctly but it doesn't map the class properties of
the command to the field used in the api itself. This will come
soon, I'm thinking of using an attribute on the fields and then
during the error conversions we can locate the attribute and use
a value defined there.

Updated all the tests, and added new ones for all the new code.