Add discovery of HTTP field name for validation
Added a new attribute that can be added to properties in commands
which will be used to determine what the HTTP field name is for
a property in a command. This is to make sure that validation
errors are returned for the field that they are sent in.

Also removed some of the reliance on traits, was going a bit mental
with them. The traits used to help build responses were just making
testing a real pain, so i've switched to using composition for this
which makes it more testable and has very little impact on the
cleanliness otherwise. If anything this seems easier to work with.