Move reponse building into commands
Moved the resonse building into the commands themselves, as it felt
less convoluted. This required building a factory to create the
response objects so that we could inject the dependencies into the
response correctly. This may still be better by just having symfony
inject a response into each command, but it feels a little odd that
way. If we decide to break response down further and have one for
each type of response, rather than the current method it would be
very painful injecting loads of responses.

Currently the response just needs a serializer and the http response
builder to make life easier, and it's quite easy to manage.

Updated all relevant tests and handlers to do this, as well as
removing the *JSONResponseBuilder classes from Connections/API.