Add a database to store connections
Updated docker-compose to include both a postgres 14 and mysql 8
database. The app should run with either database behind it, so
they are bother running to make it easier to switch between.
Added env vars for the docker compose file to define the database
configuration, which is then passed to the app.

Updated the php-fpm docker image to include all the relevant
php extensions for mysql, postgres and pdo.

Added some tasks to the makefile to make shell access to the
database containers easier. This includes both a task to start
a bash shell session in the containers, and one that will
start up a session using the containers respective cli tools
to interact with the database (either via psql or mysql).

Required and configured doctrine so that we can actually store
the data for connections in there. This includes the doctrine
bundle, with mostly standard config, but pointing at the correct
locations for the entities. I introduced some parameters for the
database credentials that are used by the doctrine config. I also
added a custom type for uuid's on the entities, configured in the
doctrine config, and the class itself in common which extends the
guid type as it shares almost all it's behaviour with it.

Configured doctrine migrations which is pretty much out of the box
except for a custom template for the generated migration files
when using the doctrine:migrations:generate console command. This
was added so that we can more easlity make the migrations database
platform agnostic, but ensures that the migrations are running on
a supported platform. This behaviour is defined by the
AbstractPlatformAwareMigration found in common.

Added a migration to create the connections table with a few fields
but more will need to be added later.

Added a Uuid generator class to separate the uuid creation from
the entities themselves. This will make things easier to test as it
can be mocked to return a predictable uuid, rather than the entity
using the static function which would unpredictable.

Updated the Connection entity to accept a uuid as part of creation,
for the same reasons mentioned above.

Added a connections repository to handle all things query related
for the connections entity. Also included an interface which is
what we typehint when we want a repository so that we can switch
it out easily at the service container if needed.

Added tests for all of the testable code. Some exceptions
are the mapping files, migration template as they are just plain
text so will be tested indirectly with intergration tests at
some point. Didn't bother unit testing the repository classes,
as it would have been alot of effort to test for very little
benefit. These will be better served by some integration tests
with an actual database, which will come later.