Add integration tests for connections repo
The connections repo need to be tested with a database.

This required setting up the basic framework for integration tests.
A new testcase is created for the integration tests which will
manage creation and deletion of databases for tests as well as the
migrations. It adds a couple of small convenience methods as well.

Tests were added for the connections repo as well as a byId method.