Add pagination and sorting for list
Added pagination and sorting options for the list connecions
command. This included some generic objects to encapsulate sorting
and pagination on queries.

Updated the connections repo to handle this including a default if
not pagination is supplied that the all method will get all rows
from the table, but will get them in batches.

Added/updated all relevant tests.